Extracted adventure drawing.
